refactor: create radix-vue/date plugin
