+++ packages/plugins/package.json
@@ -18,6 +18,7 @@
     "stub": "unbuild --stub"
   },
   "dependencies": {
+    "@internationalized/date": "^3.5.2",
     "@nuxt/kit": "npm:@nuxt/kit@latest",
     "@nuxt/schema": "latest",
     "radix-vue": "file:../radix-vue",
