+++ packages/radix-vue/constant/components.ts
@@ -401,9 +401,7 @@ export const utilities = {
     'useStateMachine',
     'useBodyScrollLock',
     'useDateFormatter',
-    'useDateUtils',
     'withDefault',
     'createContext',
-    'toDate',
   ],
 }
