+++ packages/radix-vue/date.d.ts
@@ -1,0 +1,1 @@
+export * from "./dist/date"
