+++ packages/radix-vue/package.json
@@ -31,6 +31,11 @@
       "require": "./dist/index.umd.cjs",
       "import": "./dist/index.js"
     },
+    "./date": {
+      "types": "./date.d.ts",
+      "require": "./dist/date/index.cjs",
+      "import": "./dist/date/index.mjs"
+    },
     "./namespaced": {
       "types": "./namespaced.d.ts",
       "require": "./dist/namespaced/index.cjs",
@@ -55,7 +60,8 @@
     "./dist",
     "./LICENSE",
     "./resolver.d.ts",
-    "./namespaced.d.ts"
+    "./namespaced.d.ts",
+    "./date.d.ts"
   ],
   "scripts": {
     "build": "pnpm type-check && pnpm build-only",
