+++ packages/radix-vue/src/DateField/useDateField.ts
@@ -1,5 +1,6 @@
 import { type Formatter, useKbd } from '@/shared'
-import { type AnyExceptLiteral, type HourCycle, type SegmentPart, type SegmentValueObj, getDaysInMonth, toDate } from '@/shared/date'
+import { type AnyExceptLiteral, type HourCycle, type SegmentPart, type SegmentValueObj } from '@/shared/date'
+import { getDaysInMonth, toDate } from 'date'
 import { type CalendarDateTime, type CycleTimeOptions, type DateFields, type DateValue, type TimeFields } from '@internationalized/date'
 import { type Ref, computed } from 'vue'
 import { isAcceptableSegmentKey, isNumberString, isSegmentNavigationKey } from './utils'
