+++ packages/radix-vue/src/DateRangeField/DateRangeFieldRoot.vue
@@ -8,14 +8,16 @@ import {
   type DateRange,
   type Granularity,
   type HourCycle,
-  type Matcher,
   type SegmentPart,
   type SegmentValueObj,
-  areAllDaysBetweenValid,
   getDefaultDate,
+} from '@/shared/date'
+import {
+  type Matcher,
+  areAllDaysBetweenValid,
   hasTime,
   isBefore,
-} from '@/shared/date'
+} from 'date'
 import { createContent, initializeSegmentValues, isSegmentNavigationKey, syncSegmentValues } from '@/DateField/utils'
 import type { Direction } from '@/shared/types'
