+++ packages/radix-vue/src/DateRangePicker/DateRangePickerRoot.vue
@@ -3,7 +3,8 @@ import { type DateValue } from '@internationalized/date'
 
 import type { Ref } from 'vue'
 import { createContext, useDirection } from '@/shared'
-import { type DateRange, type Granularity, type HourCycle, type Matcher, type WeekDayFormat, getDefaultDate } from '@/shared/date'
+import { type DateRange, type Granularity, type HourCycle, getDefaultDate } from '@/shared/date'
+import { type Matcher, type WeekDayFormat } from 'date'
 
 import { type CalendarRootProps, type DateRangeFieldRoot, type DateRangeFieldRootProps, PopoverRoot, type PopoverRootEmits, type PopoverRootProps } from '..'
 import type { Direction } from '@/shared/types'
