+++ packages/radix-vue/src/RangeCalendar/RangeCalendarRoot.vue
@@ -4,8 +4,9 @@ import { type DateValue, isEqualDay, isSameDay } from '@internationalized/date'
 import type { Ref } from 'vue'
 import type { PrimitiveProps } from '@/Primitive'
 import { type Formatter, createContext, useDirection } from '@/shared'
-import { createDecade, createYear, getDefaultDate, handleCalendarInitialFocus, isBefore } from '@/shared/date'
-import type { DateRange, Grid, Matcher, WeekDayFormat } from '@/shared/date'
+import { getDefaultDate, handleCalendarInitialFocus } from '@/shared/date'
+import { type Grid, type Matcher, type WeekDayFormat, createDecade, createYear, isBefore } from 'date'
+import type { DateRange } from '@/shared/date'
 import { useRangeCalendarState } from './useRangeCalendar'
 import { useCalendar } from '@/Calendar/useCalendar'
 import type { Direction } from '@/shared/types'
