+++ packages/radix-vue/src/shared/index.ts
@@ -4,7 +4,7 @@ export { isValidVNodeElement } from './isValidVNodeElement'
 export * from './browser'
 export * from './arrays'
 export * from './object'
-export { toDate, type DateRange } from './date'
+export { type DateRange } from './date'
 export { onFocusOutside } from './onFocusOutside'
 export { renderSlotFragments } from './renderSlotFragments'
 export { trapFocus } from './trap-focus'
@@ -12,7 +12,6 @@ export { useArrowNavigation } from './useArrowNavigation'
 export { useBodyScrollLock } from './useBodyScrollLock'
 export { useCollection } from './useCollection'
 export { useDateFormatter, type Formatter } from './useDateFormatter'
-export { useDateUtils } from './useDateUtils'
 export { useDirection } from './useDirection'
 export { useEmitAsProps } from './useEmitAsProps'
 export { useFocusGuards } from './useFocusGuards'
